Trying to support the vitalsService to check if the store is a mountpoint
Since the windowsPlatform/windowsDiskManager doesn't have a mounter, I'm
trying to find a sane way of supporting this in Linux.
