+++ platform/provider.go
@@ -71,7 +71,7 @@ func NewProvider(logger boshlog.Logger, dirProvider boshdirs.Provider, statsColl
 	// Kick of stats collection as soon as possible
 	statsCollector.StartCollecting(SigarStatsCollectionInterval, nil)
 
-	vitalsService := boshvitals.NewService(statsCollector, dirProvider)
+	linxuVitalsService := boshvitals.NewService(statsCollector, dirProvider, linuxDiskManager.GetMounter())
 
 	ipResolver := boship.NewResolver(boship.NetworkInterfaceToAddrsFunc)
 
@@ -143,7 +143,7 @@ func NewProvider(logger boshlog.Logger, dirProvider boshdirs.Provider, statsColl
 			compressor,
 			copier,
 			dirProvider,
-			vitalsService,
+			linxuVitalsService,
 			linuxCdutil,
 			linuxDiskManager,
 			centosNetManager,
@@ -167,7 +167,7 @@ func NewProvider(logger boshlog.Logger, dirProvider boshdirs.Provider, statsColl
 			compressor,
 			copier,
 			dirProvider,
-			vitalsService,
+			linxuVitalsService,
 			linuxCdutil,
 			linuxDiskManager,
 			ubuntuNetManager,
@@ -221,7 +221,7 @@ func NewProvider(logger boshlog.Logger, dirProvider boshdirs.Provider, statsColl
 			compressor,
 			copier,
 			dirProvider,
-			vitalsService,
+			linxuVitalsService,
 			linuxCdutil,
 			linuxDiskManager,
 			opensuseNetManager,
